Extract base BinaryOperation.
