+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/BinaryCoalescingOperation.java
@@ -10,19 +10,13 @@ import org.enso.table.data.column.builder.BuilderForType;
 import org.enso.table.data.column.operation.map.MapOperationProblemAggregator;
 import org.enso.table.data.column.storage.ColumnStorage;
 import org.enso.table.data.column.storage.Storage;
-import org.enso.table.data.column.storage.type.DateTimeType;
-import org.enso.table.data.column.storage.type.DateType;
-import org.enso.table.data.column.storage.type.NullType;
-import org.enso.table.data.column.storage.type.NumericType;
-import org.enso.table.data.column.storage.type.StorageType;
-import org.enso.table.data.column.storage.type.TextType;
-import org.enso.table.data.column.storage.type.TimeOfDayType;
+import org.enso.table.data.column.storage.type.*;
 import org.enso.table.data.table.Column;
 import org.enso.table.error.UnexpectedTypeException;
 import org.enso.table.problems.BlackholeProblemAggregator;
 import org.graalvm.polyglot.Value;
 
-public class BinaryCoalescingOperation<T> implements BinaryOperation<T> {
+public class BinaryCoalescingOperation<T> extends BinaryOperationBase<T> {
   private static Column applyOperation(
       Column left,
       Object right,
@@ -103,6 +97,7 @@ public class BinaryCoalescingOperation<T> implements BinaryOperation<T> {
           case DateTimeType dt -> DATE_TIME_MIN;
           case TimeOfDayType t -> TIME_MIN;
           case TextType t -> TEXT_MIN;
+          case BooleanType b -> BinaryCoalescingOperationBool.MIN_INSTANCE;
           case NumericType n -> BinaryCoalescingOperationNumeric.create(
               leftStorage.getType(), right, BinaryCoalescingOperationNumeric.MIN_OPERATION);
           default -> null;
@@ -142,6 +137,7 @@ public class BinaryCoalescingOperation<T> implements BinaryOperation<T> {
           case DateTimeType dt -> DATE_TIME_MAX;
           case TimeOfDayType t -> TIME_MAX;
           case TextType t -> TEXT_MAX;
+          case BooleanType b -> BinaryCoalescingOperationBool.MAX_INSTANCE;
           case NumericType n -> BinaryCoalescingOperationNumeric.create(
               leftStorage.getType(), right, BinaryCoalescingOperationNumeric.MAX_OPERATION);
           default -> null;
@@ -150,23 +146,11 @@ public class BinaryCoalescingOperation<T> implements BinaryOperation<T> {
         left, right, fallback, fallbackType, name, problemBuilder, operation, leftStorage);
   }
 
-  private final StorageType<T> validType;
   private final BiFunction<T, T, T> zipOperation;
 
   protected BinaryCoalescingOperation(StorageType<T> validType, BiFunction<T, T, T> zipOperation) {
-    this.validType = validType;
+    super(validType);
     this.zipOperation = zipOperation;
-  }
-
-  @Override
-  public boolean canApplyMap(ColumnStorage<?> left, Object rightValue) {
-    return validType.isOfType(left.getType());
-  }
-
-  @Override
-  public boolean canApplyZip(ColumnStorage<?> left, ColumnStorage<?> right) {
-    return canApplyMap(left, null)
-        && (NullType.INSTANCE.isOfType(right.getType()) || canApplyMap(right, null));
   }
 
   @Override
@@ -200,10 +184,5 @@ public class BinaryCoalescingOperation<T> implements BinaryOperation<T> {
         size -> makeStorageBuilder(size, left.getType(), right.getType()),
         false,
         (index, l, r) -> l == null ? r : (r == null ? l : zipOperation.apply(l, r)));
-  }
-
-  protected BuilderForType<T> makeStorageBuilder(
-      long size, StorageType<?> leftType, StorageType<?> rightType) {
-    return validType.makeBuilder(size, BlackholeProblemAggregator.INSTANCE);
   }
 }
